Refactoring: Klassen auf verschiedene Dateien verteilen Levels mit Mission
